+++ packages/terra-alert/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
+* updated package.json test scripts
+
+4.7.0 - (July 30, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 4.6.0 - (July 24, 2019)
 ------------------